+++ packages/pieces/google-drive/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-google-drive",
-  "version": "0.4.2"
+  "version": "0.5.0"
 }
