+++ packages/pieces/google-drive/src/lib/common/index.ts
@@ -4,33 +4,50 @@ import dayjs from "dayjs";
 
 export const common = {
     properties: {
-        folder: Property.Dropdown({
+        parentFolder: Property.Dropdown({
             displayName: "Parent Folder",
             required: false,
-            refreshers: [],
-            options: async ({ auth }) => {
-                if (!auth) {
-                    return {
-                        disabled: true,
-                        options: [],
-                        placeholder: 'Please authenticate first'
-                    }
-                }
-
-                const authProp: OAuth2PropertyValue = auth as OAuth2PropertyValue;
-                const folders: any = await common.getFolders(authProp);
-
+            refreshers: ['include_team_drives'],
+            options: async ({ auth,include_team_drives }) => {
+              if (!auth) {
                 return {
-                    disabled: false,
-                    options: folders.map((sheet: { id: string, name: string }) => {
-                        return {
-                            label: sheet.name,
-                            value: sheet.id
-                        }
-                    })
-                };
+                  disabled: true,
+                  options: [],
+                  placeholder: 'Please authenticate first'
+                }
+              }
+              const authProp: OAuth2PropertyValue = auth as OAuth2PropertyValue;
+              const folders = (await httpClient.sendRequest<{ files: { id: string, name: string }[] }>({
+                method: HttpMethod.GET,
+                url: `https://www.googleapis.com/drive/v3/files`,
+                queryParams: {
+                  q: "mimeType='application/vnd.google-apps.folder'",
+                  includeItemsFromAllDrives: include_team_drives ? "true" : "false",
+                  supportsAllDrives: "true"
+                },
+                authentication: {
+                  type: AuthenticationType.BEARER_TOKEN,
+                  token: authProp!['access_token'],
+                }
+              })).body.files;
+    
+              return {
+                disabled: false,
+                options: folders.map((folder: { id: string, name: string }) => {
+                  return {
+                    label: folder.name,
+                    value: folder.id
+                  }
+                })
+              };
             }
-        }),
+          }),
+          include_team_drives: Property.Checkbox({
+            displayName: 'Include Team Drives',
+            description: 'Determines if folders from Team Drives should be included in the results.',
+            defaultValue: false,
+            required: false,
+        })
     },
 
     async getFiles(auth: OAuth2PropertyValue, search?: {
@@ -81,5 +98,6 @@ export const common = {
         });
 
         return response.body.files;
-    }
+    },
+
 }
